+++ Cargo.toml
@@ -39,7 +39,6 @@ serde_derive = "~1.0"
 serde_json = "~1.0"
 derive_more = "~0.99"
 json = "~0.12"
-actix-web = "~3"
 futures = { version = "~0.3", optional = true }
 rayon = { version = "~1.5", optional = true }
 pow_sha256 = { version = "~0.2", optional = true }
@@ -60,3 +59,4 @@ features = ["blocking"]
 
 [dev-dependencies]
 actix-rt = "^1"
+actix-web = "~3"
