+++ src/dua/data_uses.rs
@@ -1,0 +1,383 @@
+use serde_json::Value;
+
+pub fn read_json_data_uses() -> Value {
+    let data = r#"[
+    {
+      "fides_key": "provide",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Provide the capability",
+      "description": "Provide, give, or make available the product, service, application or system.",
+      "parent_key": null,
+      "legal_basis": "Legitimate Interests",
+      "special_category": "Vital Interests",
+      "recipients": ["marketing team", "dog shelter"],
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": "https://example.org/legitimate_interest_assessment",
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "provide.service",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Service",
+      "description": "The source service, system, or product being provided to the user.",
+      "parent_key": "provide",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "provide.service.operations",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Service Operations",
+      "description": "Use of specified data categories to operate and protect in order to provide the service.",
+      "parent_key": "provide.service",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "provide.service.operations.support",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Operations Support",
+      "description": "Use of specified data categories to provide support for operation and protection in order to provide the service.",
+      "parent_key": "provide.service.operations",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "provide.service.operations.support.optimization",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Support Optimization",
+      "description": "Use of specified data categories to optimize and improve support operations in order to provide the service.",
+      "parent_key": "provide.service.operations.support",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "provide.service.upgrades",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Offer Upgrades",
+      "description": "Offer upgrades or upsales such as increased capacity for the service based on monitoring of service usage.",
+      "parent_key": "provide.service",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "improve",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Improve the capability",
+      "description": "Improve the product, service, application or system.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "improve.system",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "System",
+      "description": "The source system, product, service or application being improved.",
+      "parent_key": "improve",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "personalize",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Personalize the capability",
+      "description": "Personalize the product, service, application or system.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "personalize.system",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "System",
+      "description": "The source system, product, service or application being personalized.",
+      "parent_key": "personalize",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Advertising, Marketing or Promotion",
+      "description": "The promotion of products or services targeted to users based on the the processing of user provided data in the system.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising.first_party",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "First Party Advertising",
+      "description": "The promotion of products or services targeting users based on processing of derviced data from prior use of the system.",
+      "parent_key": "advertising",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising.third_party",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Third Party Advertising",
+      "description": "The promotion of products or services targeting users based on processing of specific categories of data acquired from third party sources.",
+      "parent_key": "advertising",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising.first_party.contextual",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "First Party Contextual Advertising",
+      "description": "The promotion of products or services targeted to users based on the processing of data from the users prior use of the services.",
+      "parent_key": "advertising.first_party",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising.first_party.personalized",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "First Party Personalized Advertising",
+      "description": "The targeting and changing of promotional content based on processing of specific data categories from the user.",
+      "parent_key": "advertising.first_party",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "advertising.third_party.personalized",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Third Party Personalized Advertising",
+      "description": "The targeting and changing of promotional content based on processing of specific categories of user data acquired from third party sources.",
+      "parent_key": "advertising.third_party",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "third_party_sharing",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Third Party Sharing",
+      "description": "The transfer of specified data categories to third parties outside of the system/application's scope.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "third_party_sharing.payment_processing",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Sharing for Processing Payments",
+      "description": "Sharing of specified data categories with a third party for payment processing.",
+      "parent_key": "third_party_sharing",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "third_party_sharing.personalized_advertising",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Sharing for Personalized Advertising",
+      "description": "Sharing of specified data categories for the purpose of marketing/advertising/promotion.",
+      "parent_key": "third_party_sharing",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "third_party_sharing.fraud_detection",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Sharing for Fraud Detection",
+      "description": "Sharing of specified data categories with a third party fo fraud prevention/detection.",
+      "parent_key": "third_party_sharing",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "third_party_sharing.legal_obligation",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Sharing for Legal Obligation",
+      "description": "Sharing of data for legal obligations, including contracts, applicable laws or regulations.",
+      "parent_key": "third_party_sharing",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "collect",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Collect",
+      "description": "Collecting and storing data in order to use it for another purpose such as data training for ML.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "train_ai_system",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Train AI System",
+      "description": "Training an AI system. Please note when this data use is specified, the method and degree to which a user may be directly identified in the resulting AI system should be appended.",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": true,
+      "active": true
+    },
+    {
+      "fides_key": "custom",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Custom field",
+      "description": "Custom field for tests",
+      "parent_key": null,
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": false,
+      "active": true
+    },
+    {
+      "fides_key": "custom.foo",
+      "organization_fides_key": "default_organization",
+      "tags": null,
+      "name": "Custom foo",
+      "description": "Custom foo for tests",
+      "parent_key": "custom",
+      "legal_basis": null,
+      "special_category": null,
+      "recipients": null,
+      "legitimate_interest": false,
+      "legitimate_interest_impact_assessment": null,
+      "is_default": false,
+      "active": true
+    }
+  ]"#;
+    let v: Value = serde_json::from_str(data).unwrap();
+    v
+}
